requirements.txt
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-sslserver==0.21
djangorestframework==3.10.3
h5py==2.10.0
idna==2.8
Markdown==3.1.1
numpy==1.17.3
oauthlib==3.1.0
Pillow==6.2.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.6
